Adding tests for data races
