+++ tests/CMakeLists.txt
@@ -23,11 +23,17 @@ add_c_test(add_offset)
 find_package(Threads)
 if(Threads_FOUND)
   message(STATUS "Your system supports threads.")
-  if(ROARING_BUILD_C_TESTS_AS_CPP)  # under C++, container_t* != void*
-    SET_SOURCE_FILES_PROPERTIES(threads_unit.c PROPERTIES LANGUAGE CXX)
-  endif()
-  add_executable(threads_unit threads_unit.c)
+  add_executable(threads_unit threads_unit.cpp)
   target_link_libraries(threads_unit PRIVATE roaring Threads::Threads)
+  if(ROARING_SANITIZE_THREADS)
+    # libtsan might be needed
+    if(CMAKE_SYSTEM_NAME STREQUAL "Linux")
+      message(STATUS "Under Linux, you may need to install libtsan." )
+    endif()
+    target_compile_options(threads_unit PRIVATE -fsanitize=thread -fno-sanitize-recover=all)
+    target_link_options(threads_unit PRIVATE -fsanitize=thread -fno-sanitize-recover=all)
+    message(STATUS "Sanitizing threads.")
+  endif()
   add_test(threads_unit threads_unit)
 else(Threads_FOUND)
   message(STATUS "Your system does not support threads.")
