--- tests/threads_unit.c
@@ -1,84 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-
-#if __STDC_NO_THREADS__
-int main() {
-    printf("This test requires threads support.\n");
-    return EXIT_SUCCESS;
-}
-#else //__STDC_NO_THREADS__
-#include <threads.h>
-#include <roaring/roaring.h>
-#include <roaring/misc/configreport.h>
-
-// helper function to create a roaring bitmap from an array
-static roaring_bitmap_t *make_roaring_from_array(uint32_t *a, int len) {
-    roaring_bitmap_t *r1 = roaring_bitmap_create();
-    for (int i = 0; i < len; ++i) roaring_bitmap_add(r1, a[i]);
-    return r1;
-}
-
-// We are mostly running this test to check for data races suing thread sanitizer.
-int run(void * input) {
-    roaring_bitmap_t **rarray = (roaring_bitmap_t **)input;
-    roaring_bitmap_t *r1 = roaring_bitmap_copy(rarray[0]);
-    roaring_bitmap_t *r2 = rarray[1];
-    roaring_bitmap_t *r3 = rarray[2];
-    roaring_bitmap_and_inplace(r1, r2);
-    roaring_bitmap_and_inplace(r1, r3);
-    int answer = (int)roaring_bitmap_get_cardinality(r1);
-    roaring_bitmap_free(r1);
-    return answer;
-}
-
-bool run_threads_unit_tests() {
-    uint32_t *ans = (uint32_t*)calloc(100000, sizeof(int32_t));
-
-    for (uint32_t i = 0; i < 50000; i++) {
-        if (i != 300) {
-            ans[ans_ctr++] = 65536 + i;
-        }
-    }
-    for (uint32_t i = 50000; i < 150000; i++) {
-        if ((i%500) == 0) {
-            ans[ans_ctr++] = i;
-        }
-    }
-    for (uint32_t i = 150000; i < 200000; i++) {
-        if ((i%2) == 0) {
-            ans[ans_ctr++] = i;
-        }
-    }
-    
-    roaring_bitmap_t *r1 = make_roaring_from_array(ans, ans_ctr);
-    free(arr);
-
-    roaring_bitmap_set_copy_on_write(r1, true);
-    roaring_bitmap_run_optimize(r1);
-    roaring_bitmap_t *r2 = roaring_bitmap_of(5, 10010,10020,10030,10040,10050);
-    roaring_bitmap_set_copy_on_write(r2, true);
-    roaring_bitmap_t *r3 = roaring_bitmap_copy(r1);
-    roaring_bitmap_t* rarray[] = {r1, r2, r3};
-
-    thrd_t thread1;
-    thrd_t thread2;
-    int result1{};
-    int result2{};
-
-    thrd_create(&thread1, run, rarray);
-    thrd_create(&thread2, run, rarray);
-    thrd_join(&thread1, &result1);
-    thrd_join(&thread2, &result2);
-    roaring_bitmap_free(r1);
-    roaring_bitmap_free(r2);
-    roaring_bitmap_free(r3);
-    return result1 == result2;
-}
-
-int main() {
-    tellmeall();
-    bool is_ok = run_threads_unit_tests();
-    return is_ok: EXIT_SUCCESS ? EXIT_FAILURE;
-}
-
-#endif // __STDC_NO_THREADS__
