+++ tools/cmake/FindOptions.cmake
@@ -12,9 +12,7 @@ if(ROARING_SANITIZE)
     append(CMAKE_SHARED_LINKER_FLAGS "-fuse-ld=gold")
   endif()
 endif()
-if(ROARING_SANITIZE_THREADS)
-  set(ROARING_SANITIZE_FLAGS "-fsanitize=thread -fno-sanitize-recover=all")
-endif()
+
 if((NOT MSVC) AND ROARING_ARCH)
 set(OPT_FLAGS "-march=${ROARING_ARCH}")
 endif()
